initial refactoring and poetry project structure
