+++ patmat/io/file_utils.py
@@ -1,0 +1,13 @@
+import gzip
+import bz2
+def openfile(file):
+    '''
+    Opens a file
+    '''
+    if file.endswith('.gz'):
+        opened_file = gzip.open(file,'rt')
+    elif file.endswith('bz') or file.endswith('bz2'):
+        opened_file = bz2.open(file,'rt')
+    else:
+        opened_file = open(file,'rt')
+    return opened_file
